backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
marshmallow==3.20.1
bcrypt==4.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7